+++ bee-common/bee-common/src/logger/config.rs
@@ -20,8 +20,10 @@ const DEFAULT_OUTPUT_LEVEL: LevelFilter = LevelFilter::Info;
 pub struct LoggerOutputConfigBuilder {
     /// Name of an output file, or `stdout` for standard output.
     name: Option<String>,
-    /// Log level of an output.
-    level: Option<LevelFilter>,
+    /// Log level filter of an output.
+    level_filter: Option<LevelFilter>,
+    /// Log target filters of an output.
+    target_filters: Option<Vec<String>>,
 }
 
 impl LoggerOutputConfigBuilder {
@@ -37,8 +39,15 @@ impl LoggerOutputConfigBuilder {
     }
 
     /// Sets the level of a logger output.
-    pub fn level(mut self, level: LevelFilter) -> Self {
-        self.level.replace(level);
+    pub fn level_filter(mut self, level: LevelFilter) -> Self {
+        self.level_filter.replace(level);
+        self
+    }
+
+    /// Sets a collection of filters of a logger output.
+    /// A message is logged only if one of the filters is part of the log's metadata target.
+    pub fn target_filters(mut self, target_filters: &[&str]) -> Self {
+        self.target_filters = Some(target_filters.iter().map(|f| f.to_string()).collect::<Vec<String>>());
         self
     }
 
@@ -46,7 +55,13 @@ impl LoggerOutputConfigBuilder {
     pub fn finish(self) -> LoggerOutputConfig {
         LoggerOutputConfig {
             name: self.name.unwrap_or_else(|| DEFAULT_OUTPUT_NAME.to_owned()),
-            level: self.level.unwrap_or(DEFAULT_OUTPUT_LEVEL),
+            level_filter: self.level_filter.unwrap_or(DEFAULT_OUTPUT_LEVEL),
+            target_filters: self
+                .target_filters
+                .unwrap_or_else(Vec::new)
+                .iter()
+                .map(|f| f.to_lowercase())
+                .collect(),
         }
     }
 }
@@ -57,7 +72,9 @@ pub struct LoggerOutputConfig {
     /// Name of an output file, or `stdout` for standard output.
     pub(crate) name: String,
     /// Log level of an output.
-    pub(crate) level: LevelFilter,
+    pub(crate) level_filter: LevelFilter,
+    /// Log filters of the output.
+    pub(crate) target_filters: Vec<String>,
 }
 
 /// Builder for a logger configuration.
@@ -85,7 +102,7 @@ impl LoggerConfigBuilder {
                 Some(output_name) => output_name[..] == name,
                 None => false,
             }) {
-                stdout.level.replace(level);
+                stdout.level_filter.replace(level);
             }
         }
     }