+++ bee-common/bee-common/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `Packable` implementation for `Vec<T: Packable>`
+- Add `filters` option to the logger.
 
 ### Changed